refactor: AppWrapper creates the appContext and provides it through a Context
